LayoutTests/imported/w3c/web-platform-tests/shadow-dom/slots-outside-shadow-dom-expected.txt

foo

FAIL Light DOM slot element should be in flattened assignedNodes assert_array_equals: property 0, expected Element node <slot>foo</slot> but got Text node "foo"
